+++ eslint.config.js
@@ -11,6 +11,7 @@ export default tseslint.config(
       "packages/d6-system-2e-core-content/d6-system-2e-core-content.mjs",
       "packages/d6-system-2e-fantasy/d6-system-2e-fantasy.mjs",
       "packages/d6-system-2e-science-fiction/d6-system-2e-science-fiction.mjs",
+      "packages/d6-system-2e-superhero/d6-system-2e-superhero.mjs",
       "packages/open-d6-space-d6-system-2e/open-d6-space-d6-system-2e.mjs",
       "packages/open-d6-fantasy-d6-system-2e/open-d6-fantasy-d6-system-2e.mjs",
       "packages/open-d6-fantasy-d6-system-2e/content/catalog.d.mts",