+++ packages/d6-system-2e-science-fiction/d6-system-2e-science-fiction.mjs
@@ -16,7 +16,7 @@ Hooks.once("ready", () => {
     mechanicIds: ["science-fiction-skills"],
     recommendedPrimaryProfile: "second-edition",
     rulesFamily: "d6-system-second-edition",
-    version: "0.1.0-alpha.27"
+    version: "0.1.0-alpha.28"
   });
 });
 //# sourceMappingURL=d6-system-2e-science-fiction.mjs.map