+++ scripts/build.mjs
@@ -31,6 +31,14 @@ const secondEditionScienceFictionOutput = path.join(
   secondEditionScienceFictionDirectory,
   "d6-system-2e-science-fiction.mjs",
 );
+const secondEditionSuperheroDirectory = path.join(
+  root,
+  "packages/d6-system-2e-superhero",
+);
+const secondEditionSuperheroOutput = path.join(
+  secondEditionSuperheroDirectory,
+  "d6-system-2e-superhero.mjs",
+);
 const hudOutput = path.join(hudDirectory, "token-action-hud-d6-system-2e.mjs");
 const spaceDirectory = path.join(root, "packages/open-d6-space-d6-system-2e");
 const spaceOutput = path.join(spaceDirectory, "open-d6-space-d6-system-2e.mjs");
@@ -56,6 +64,8 @@ const outputs = [
   `${secondEditionFantasyOutput}.map`,
   secondEditionScienceFictionOutput,
   `${secondEditionScienceFictionOutput}.map`,
+  secondEditionSuperheroOutput,
+  `${secondEditionSuperheroOutput}.map`,
   hudOutput,
   `${hudOutput}.map`,
   spaceOutput,
@@ -75,6 +85,16 @@ await clean();
 if (!process.argv.includes("--clean")) {
   await mkdir(outputDirectory, { recursive: true });
   await Promise.all([
+    build({
+      bundle: true,
+      entryPoints: [path.join(secondEditionSuperheroDirectory, "src/main.ts")],
+      format: "esm",
+      logLevel: "info",
+      outfile: secondEditionSuperheroOutput,
+      platform: "browser",
+      sourcemap: true,
+      target: "es2022",
+    }),
     build({
       bundle: true,
       entryPoints: [